drop main dep from build_environment/subprocess_context
